+++ backend/alembic/versions/0002_create_channels.py
@@ -18,7 +18,7 @@ depends_on: Union[str, Sequence[str], None] = None
 def upgrade() -> None:
     op.create_table(
         "channels",
-        sa.Column("id", sa.Integer, primary_key=True, autoincrement=True),
+        sa.Column("id", sa.BigInteger, primary_key=True, autoincrement=True),
         sa.Column("device_id", sa.Integer, sa.ForeignKey("devices.id", ondelete="CASCADE"), nullable=False),
         sa.Column("index", sa.Integer, nullable=False),  # канал внутри устройства
         sa.Column("type", sa.String, nullable=False),    # DO, DI, AO ...
